src/census-data/pyproject.toml
[project]
name = "us-data-mcp.census-data"
version = "0.1.0"
description = "US Census Bureau data MCP server"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["mcp", "census", "demographics", "us-government"]

dependencies = [
    "mcp>=0.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
census-data-server = "us_data_mcp.census_data.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/us_data_mcp"]
